Refactor: 쿼리 전용 repository 분리
